tests/wpt/web-platform-tests/css/css-backgrounds/parsing/border-color-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-color with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-color">
<meta name="assert" content="border-color supports the full grammar '<color>{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
// Edge serializes as "currentColor".
test_valid_value("border-color", "currentcolor", ["currentcolor", "currentColor"]);
test_valid_value("border-color", "currentColor", ["currentcolor", "currentColor"]);

test_valid_value("border-color", "red yellow green blue");

test_valid_value("border-top-color", "red");
test_valid_value("border-right-color", "yellow");
test_valid_value("border-bottom-color", "green");
test_valid_value("border-left-color", "blue");
</script>
</body>
</html>
